Note about scope for a DB constraint
n.b., It would be much better if we had a table of e-mail addresses (at
which point, it would be trivial to enforce uniqueness), with a
one-to-many relationship from users. However, we're pushed for time and
so we swallow the technical debt under the assumption that a Sanger and
personal e-mail will probably be fine.