Tidy up email templates
The library the frontend uses to present a WYSIWYG editor for templates
can only cope with a very limited number of HTML tags; any unsupported
tags are removed from the output, and any text not in a tag is hoisted
to the start of the text. Therefore, we now use only a very small number
of tags (<br>, <h3>, and <p>), and ensure to include all text in the
template in tags, even though this will result in the final email
technically containing invalid HTML (for example nested <p> tags).