Schedule deadlines and send mail on rotation creation
Previously, mail (i.e. supervisor invitations) would not be sent and
deadlines would not be scheduled on rotation creation, only when the
rotation was edited (and even then, only when the date of the deadline
varied from what it was originally).

It's unclear exactly why this was broken, given that in the past it
apparently worked (and it would seem an unusual thing not to test), but
presumably during the migration from the non-REST to the REST API, this
section of code was missed.