Allow students to be pre-assigned to projects by supervisors
This is only the backend piece of the puzzle; the frontend UI and code
to send the student ID still needs to be written.

This also currently allows a single student to be assigned to multiple
projects in the same rotation, which is *bad* (although it seems like
this would probably be fixed in the finalisation process). IOW, there is
still work to do.