Improve async test tooling
Keeping the event loop around for the duration of each testcase may help
improve performance, and the AsyncTestMeta metaclass should make writing
async tests a little less error-prone and boilerplatey.

Further work: Detecting cases where an async function has been decorated
with an async-unaware decorator, like unittest.mock.patch, and
automatically inserting the async_test shim between the original test
function and the decorator. (__wrapped__ is one piece of the puzzle here
(if the decorator doesn't set that, it's broken and we have no hope of
rescuing the situation), but the decorator still needs to be reapplied
to async_test(__wrapped__), which probably requires some exciting
introspection. Also consider: how should we tell the difference between
an async-aware and an async-unaware decorator? Would it be easier to
define async-aware versions of the common decorators instead?)