Set pre-assigned projects to be the student's first choice
This avoids the case where, for example, a student arranges to do a
project which is then pre-assigned to them, but doesn't make any choices
for their preferred projects, resulting in the student being assigned an
inflated priority once projects are finalised.

(Of course, students can still game the system by selecting their
pre-assigned project as their third choice after it is assigned to them;
perhaps it would be better to prevent students from changing their
choices once they have been pre-assigned a project? Or, require that all
choices are contiguous -- that is, you can't make a second choice until
you have made a first choice? That probably results in a significantly
worse UX when rearranging choices, though...)