Correct MIME type of exported Excel spreadsheets
We are generating xlsx files, not xls files, so this MIME type is more
correct.

In addition, an issue was observed during user testing where Firefox on
Windows 10 appears to change the file extension based on the declared
MIME type of the content, so the file would end up with the extension
.xls anyway -- this then caused a warning in Excel that the file
extension did not match the contents.