Add sensible default invitation templates
Previously there were no default templates for the supervisor_invite_*
or student_invite_* emails.

Since the supervisor_invite_* templates must handle extensions and
manual reminders, the interface for that template has changed slightly
(previously a `new_deadline` or `extension` could be passed instead of
or in addition to the `deadline`; now the template takes a `rotation`,
and optionally an `old_deadline` and/or a `reminder` flag).

Although the emails for the three rotations can differ, it's fine for
them to default to the same email, so the _2 and _3 templates are
symlinked to the _1 template.

Fixes #8, closes #9.