Avoid printing a scary traceback when an email template is broken
Uses some janky introspection of the traceback to still return some
useful debugging information.

Also changes the way templates are constructed, so that they always use
the same Jinja2 Environment (so that custom filters, which are checked
statically, are always available).

This isn't perfect; the content currently has to be placed inside a
<pre> on the client so that the line breaks render properly, for
example.