Use one job with multiple triggers for reminders
Previously, one job would be scheduled for each "pester point" of a
deadline. Combined with the very long misfire_grace_time, this would
result in several reminders being sent simultaneously if they were
scheduled in the past, and in that case the notification would also
state the wrong number of days left (since the delta was hardcoded).

Now, each deadline's reminders are contained within a single job, which
is triggered at several predefined dates relative to the deadline. The
offset between the job and the deadline is calculated when the reminder
is sent, and if two reminders are scheduled to run simultaneously (e.g.
the 14-day and 7-day reminders), only the most recent reminder will be
sent (see #12).

Fixes #7.