Add API endpoint to allow students to remind themselves of their votes
Note that the atrocities in the template are /required/ for the frontend
editor not to choke when users try to edit it; for example,
semantically-correct lists are impossible, because the editor will not
understand the for-loop syntax between the <ol> and <li> elements.
Hopefully the user's email client will collapse empty <p> tags,
otherwise they will see a lot of whitespace everywhere -- some of this
could be removed by putting more Jinja2 tags on one line, but that then
makes the template significantly more dense and less readable for the
Graduate Office.

Also, using the characters "<" and ">" in the templates is not allowed,
as the editor will convert them to "&lt;" and "&gt;" respectively, which
Jinja2 (understandably) barfs on, so we have to use the long-form
"greaterthan" test instead. The "reject" filter also produces a
generator, so we have to convert it to a list before taking its length.

See wsi-cogs/frontend#9.