Refactor calling convention for scheduled jobs
Jobs are now scheduled using keyword arguments, which means that they
all implement the new _Job protocol -- in the future, this will allow
them to be more thoroughly typechecked.

Several jobs took optional arguments that were never passed by callers:
- reminder no longer takes a list of users to remind
- reminder template is no longer looked up based on the rotation

In addition, the rotation ID is now stored along with the scheduled job,
which will be required if the rest of the system is ever updated to
support multiple simultaneous rotations.

**This commit will break if deployed onto an existing instance with jobs
scheduled!**

Fixes #6.