Do not let students upload projects until they are finalised
Since commit 27c2d10 ("Fix bug where upload project button wouldn't
appear"), students have been able to upload projects as soon as the
following conditions are both true:

- the deadline for choosing projects (student_choice) has passed
- the Graduate Office (or a supervisor) has assigned the student a
  project

This is the case even before the Graduate Office has pressed the
"Finalise" button, and before the emails announcing which students have
been assigned to which projects are sent out. Therefore, this seems
incredibly likely to be undesired behaviour.

In any case, it's not possible that students would have any work to
upload before their projects are finalised, so as long as the rotation
is set to uploadable once projects are finalised (which it currently
is), this should have no negative effect.