Add setuptools machinery
Unsolved problem: how to do deployment? Using requirements.txt as a
constraints file means rebuilding the entire Docker image every time a
file is changed, which is no good. Just using it as a requirements file
probably makes more sense, but that can't happen until there's a
configuration option to enable/disable PageSmith support (instead of
enabling it based on whether or not the dependencies are available).