Correctly pluralise "was" in late submission reminder
This also replaces the dependency on natural with a dependency on
inflect, which appears to be more actively-maintained.

It's unfortunate that this makes the template even harder to read, but
there's no obvious way to improve the situation (apart from rewording
the template to remove dependencies on the number of files to be
submitted).