Add notification for student_complete change of deadline
This sucks; unfortunately I can't immediately think of a better
abstraction that wouldn't be significantly more work to implement.

In the long run, it might be worth trying to formalise the email
templates a little more -- by making explicit which variables are
expected by each template, maybe a better way to substitute them into
the template would become apparent.