Use NamedTuple for notification configuration
This allows us to work around the fact that mypy doesn't use refinements
made to one variable to refine other variables (e.g. previously,
refining `template` wouldn't refine `permission`).

The callable signature is a bit vague; it could be rewritten as a
protocol, like _PesterProtocol, but it's unclear that it would provide
any significant benefit. If the email templates are ever made
type-checkable, then maybe this should be revisited.