Add notification for marking_complete change of deadline
This commit switches change-of-deadline notifications to all use a
single new template, `change_of_deadline`, because retrofitting the
existing templates for this purpose was too difficult. In particular,
`student_uploaded` is potentially sent to each supervisor/CoGS member
multiple times, with different (project-specific) text each time, but
there's no need to send multiple deadline change notifications to any
one user.

If a rotation is created far in advance, and the dates are changed after
it is created, then users could be emailed about a rotation they have
not been invited to yet -- for example, CoGS members could be emailed to
tell them that the date by which they must submit feedback has changed
before the students have even started looking at the projects. I don't
see an easy solution for this.