Include libfaketime in the development Docker image
This can be used to modify the apparent flow of time inside the
container, allowing for easier and faster testing of scheduled events.

Configuration can be done by writing files into the image (either by
hand or via bindmount/volume), so it's not necessary to include
configuration in this repository. By default, libfaketime will report
the system time.