Add /api/util/time endpoint to show the current time
This is helpful when using libfaketime to adjust the application's
perception of time. As it's not used or useful outside of development,
it is disabled when the Python interpreter is not in development mode.

The three different values shown are somewhat arbitrary; the first,
`datetime.now()`, prints a human-readable representation of the time,
which should be (almost) the same as `time.time()`, which is a Unix
timestamp. The format of the event loop's `.time()` is not documented,
but in CPython 3.7 it appears to be a Unix timestamp much like
`time.time()` -- therefore `time.time()` was included to allow easier
comparison between the event loop's idea of time and that of the rest of
the application, if they ever differ (since `datetime.now()` and
`time.time()` are guaranteed to represent essentially the same time).