Make job IDs more descriptive
The table which an ID refers to is now included in the job ID. Since no
caller of schedule_deadline was using the suffix argument, it has been
removed (since the rotation series & part are included in the job ID
automatically, it's unclear what it would have been used for).

There's no documentation on what characters APScheduler considers valid
in a job ID, but with the Postgres/SQLAlchemy backend a column of type
VARCHAR(191) is used, so certainly any printable ASCII character should
be acceptable.

This commit should not be applied to an instance with scheduled jobs, as
attempting to reschedule those jobs would instead schedule new ones.