Remove strange marking deadline calculation
It looks like this was written when the grace period was a fixed amount
of time after the project was first uploaded, so the typical case would
have been for it to run before student_complete. It looked at the amount
of time that supervisors had been given to mark the projects (the time
between the deadline for students to submit and the deadline for
supervisors to mark), and gave supervisors that much time before sending
them another reminder.

Now that the grace period is a fixed amount of time after the deadline
for everyone to upload, this complicated calculation doesn't make sense.
It still works, but it's over-complicated, and can be written in a
simpler way. The net result is that supervisors will still be emailed
immediately once the grace period is over, and then next emailed a fixed
amount of time after the marking deadline, if they're yet to submit a
mark. They will then be emailed at intervals after that until they
submit a mark.