Make explicit the logic around missed upload deadline
When a student failed to upload within the grace period, previously
nothing special would happen (apart from an email notification to them
at the point they missed the deadline); in particular, they would still
be able to upload. Once they uploaded a report, the grace period job
would be scheduled for three days after the deadline -- if this was in
the past, APScheduler would determine that it had failed to execute the
job on time, and run it ASAP (i.e. as soon as it was scheduled). This
would result in their supervisor being notified immediately that they
had uploaded their project, and some suboptimal behaviour on the
frontend (since it can't cope with a project's grace period immediately
having passed).

Since bd959c3 ("Make schedule_user_deadline take and fix a date"), the
grace period job for late submissions would instead run at the next
11:59pm after the project was uploaded, which is an improvement, but in
the event of an extremely late upload or if APScheduler's maximum
misfire time was reduced, it's possible that the grace period job
wouldn't be run at all. Explicitly checking for and handling late
submissions makes the intended behaviour clearer and reduces the number
of possible edge cases.