Improve wording of error message shown to early markers
This doesn't fix the problem* of people trying to find the button to
mark a project as soon as the deadline for students to upload has
passed, before the grace period has expired, but it certainly makes the
code more obvious.

Also fix an ugly (and arguably incorrect) test -- `is not True` is not
exactly idiomatic Python.

*this was not observed or mentioned in user testing, so it may not
actually be a problem at all