Allow the Graduate Office to create projects
The Graduate Office needs to be able to perform tasks on behalf of other
users, in case those users are unavailable. This commit allows the
Graduate Office to create projects on behalf of other supervisors.

When a supervisor submits a project, they are required to be its
supervisor. Members of the Graduate Office are allowed to assign any
user to be a project's supervisor (the frontend only allows users
designated as supervisors to be chosen, and it's probably not worth
checking on the backend that the chosen user is actually a supervisor*).

This commit requires the frontend to be updated to include an additional
`supervisor` parameter in the request when creating a new project,
otherwise the request will fail.

See wsi-cogs/frontend#19.

*since a user with the `modify_permissions` permission could just change
the permissions of a user to "supervisor", create the project, then
change the permissions back.