Use strings rather than numbers for grades in the API
This removes a bit of indirection in the frontend at the cost of a
little more indirection in the backend, but more importantly it will
make it easier to remove the duplication of grade definitions in the
frontend and backend in the future.

The serialised ProjectGrades were already using strings, it's just the
/api/projects/:id/mark endpoint that was accepting a numerical ID.

This change must be deployed at the same time as the accompanying
frontend change ("Use ProjectFeedbackForm for both reading and writing
feedback"), for obvious reasons.