Allow the Graduate Office to mark project reports
The Graduate Office needs to be able to perform tasks on behalf of other
users, in case those users are unavailable. This commit allows the
Graduate Office to mark project reports on behalf of other supervisors.

This commit requires the frontend to be updated to include an additional
`marker` parameter in the request when marking a project, otherwise the
request will fail.

See wsi-cogs/frontend#19.