Make can_view_group refer to projects within the rotation
Previously, if a user was not allowed to view a rotation (per
can_view_group), then they were forbidden from receiving any information
about the rotation at all. This required a bunch of hacks on the
frontend to cope with the HTTP errors this resulted in, and meant that
the student-visible UI was sometimes pretty poor.

Changing this to only hide the projects in a rotation (rather than the
rotation itself) means that the frontend can work without totally
falling over if the latest rotation is not ready for students to view
yet, which is important if rotations are going to overlap successfully.