Add upload information (in particular, file names) to the response of API projects.upload
The names of the uploaded files are stored on the client side as part of application state. Earlier, the new files were uploaded to overwrite the old ones, the state wasn't being updated. The upload information in the response can be used to update state.

Related: Issue#31 in wsi-cogs/frontend