Fix choices shown in Finalize Student Projects screen after new choices are saved
Earlier, when a change was made to the choices, and "Save" button was clicked, the shown choices would revert to the earlier saved choices (if any). This was because the API response to assign_projects sent only the projects which were now assigned, and not ALL the projects. The API response would thus update the assignments that are saved, and not update the state of projects which dont have a student assigned to them. The fix involves sending the updated state for ALL the projects, not just the assigned one.